respect Bundle standard: move files to particular dirs
