--- AutowiringException.php
@@ -1,6 +1,0 @@
-<?php
-namespace Skrz\Bundle\AutowiringBundle;
-
-class AutowiringException extends \LogicException
-{
-}
