+++ DependencyInjection/ClassMultiMap.php
@@ -1,5 +1,9 @@
 <?php
-namespace Skrz\Bundle\AutowiringBundle;
+namespace Skrz\Bundle\AutowiringBundle\DependencyInjection;
+
+use ReflectionClass;
+use Skrz\Bundle\AutowiringBundle\Exception\MultipleValuesException;
+use Skrz\Bundle\AutowiringBundle\Exception\NoValueException;
 
 /**
  * Maps from class name, all its parents, and implemented interfaces to certain value
@@ -10,11 +14,15 @@ class ClassMultiMap
 {
 
 	/** @var array */
-	private $classes = array();
+  	private $classes = array();
 
+	/**
+	 * @param string $className
+	 * @param string $value
+	 */
 	public function put($className, $value)
 	{
-		$rc = new \ReflectionClass($className);
+		$rc = new ReflectionClass($className);
 
 		foreach ($rc->getInterfaceNames() as $interfaceName) {
 			if (!isset($this->classes[$interfaceName])) {
@@ -31,6 +39,10 @@ class ClassMultiMap
 		} while ($rc = $rc->getParentClass());
 	}
 
+	/**
+	 * @param string $className
+	 * @return string
+	 */
 	public function getSingle($className)
 	{
 		if (!isset($this->classes[$className])) {
@@ -46,6 +58,10 @@ class ClassMultiMap
 		return reset($values);
 	}
 
+	/**
+	 * @param string $className
+	 * @return string[]
+	 */
 	public function getMulti($className)
 	{
 		if (!isset($this->classes[$className])) {
