+++ Exception/AutowiringException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Skrz\Bundle\AutowiringBundle\Exception;
+
+use LogicException;
+
+class AutowiringException extends LogicException
+{
+}
