+++ Exception/MultipleValuesException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Skrz\Bundle\AutowiringBundle\Exception;
+
+use LogicException;
+
+class MultipleValuesException extends LogicException
+{
+}
