+++ Exception/NoValueException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Skrz\Bundle\AutowiringBundle\Exception;
+
+use LogicException;
+
+class NoValueException extends LogicException
+{
+
+}
