--- NoValueException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Skrz\Bundle\AutowiringBundle;
-
-class NoValueException extends \LogicException
-{
-
-}
-
